+++ R/download-table.R
@@ -26,7 +26,7 @@ download_table <- function( id
                           , cache = FALSE
                           , verbose = TRUE
                           , typed = FALSE
-                          , base_url = CBSOPENDATA){
+                          , base_url = getOption("cbsodataR.base_url", BASE_URL)){
   .Deprecated("cbs_download_table")
   cbs_download_table(id = id, ..., dir =dir, cache =cache, verbose = verbose, typed = typed, base_url = base_url)
 }